// migrations/snapshots/d16cc5a026e2ca03dc8926749f90194dd1b957f3fefbaab6ed999c3109b51efb/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Lodge: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly roomType: 'SELF_CONTAIN' | 'SINGLE_ROOM' | 'FLAT';
      readonly available_rooms: CodecTypes['pg/int4@1']['output'];
      readonly annual_rent: CodecTypes['pg/float8@1']['output'];
      readonly distance: CodecTypes['pg/float8@1']['output'];
      readonly landlord: CodecTypes['pg/text@1']['output'];
      readonly amenities: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly images: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
    };
    readonly RefreshToken: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly firstname: CodecTypes['pg/text@1']['output'];
      readonly lastname: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly role: 'STUDENT' | 'LODGE_OWNER' | 'ADMIN';
      readonly status: 'PENDING' | 'APPROVED' | 'SUSPENDED';
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Lodge: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly roomType: 'SELF_CONTAIN' | 'SINGLE_ROOM' | 'FLAT';
      readonly available_rooms: CodecTypes['pg/int4@1']['input'];
      readonly annual_rent: CodecTypes['pg/float8@1']['input'];
      readonly distance: CodecTypes['pg/float8@1']['input'];
      readonly landlord: CodecTypes['pg/text@1']['input'];
      readonly amenities: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly images: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
    };
    readonly RefreshToken: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly firstname: CodecTypes['pg/text@1']['input'];
      readonly lastname: CodecTypes['pg/text@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly role: 'STUDENT' | 'LODGE_OWNER' | 'ADMIN';
      readonly status: 'PENDING' | 'APPROVED' | 'SUSPENDED';
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly lodge: {
      readonly amenities: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly annual_rent: CodecTypes['pg/float8@1']['output'];
      readonly available_rooms: CodecTypes['pg/int4@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly distance: CodecTypes['pg/float8@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly images: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly landlord: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly roomType: 'SELF_CONTAIN' | 'SINGLE_ROOM' | 'FLAT';
    };
    readonly refreshToken: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly user: {
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly firstname: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly lastname: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly role: 'STUDENT' | 'LODGE_OWNER' | 'ADMIN';
      readonly status: 'PENDING' | 'APPROVED' | 'SUSPENDED';
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly lodge: {
      readonly amenities: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly annual_rent: CodecTypes['pg/float8@1']['input'];
      readonly available_rooms: CodecTypes['pg/int4@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly distance: CodecTypes['pg/float8@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly images: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly landlord: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly roomType: 'SELF_CONTAIN' | 'SINGLE_ROOM' | 'FLAT';
    };
    readonly refreshToken: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly user: {
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly firstname: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly lastname: CodecTypes['pg/text@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly role: 'STUDENT' | 'LODGE_OWNER' | 'ADMIN';
      readonly status: 'PENDING' | 'APPROVED' | 'SUSPENDED';
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;